refactor: 벨리데이션, 비밀번호 토글 기능 리팩토링 및 html에서 js 임포트 방법 수정 (defer)
